+++ include/minocore/dist.h
@@ -1,0 +1,5 @@
+#ifndef FGC_DISTANCE_HEADERS_
+#define FGC_DISTANCE_HEADERS_
+#include <minocore/dist/applicator.h>
+#include <minocore/dist/distance.h>
+#endif
