+++ include/minocore/minocore.h
@@ -1,30 +1,12 @@
 #ifndef FGC_ALL_INCLUDE__
 #define FGC_ALL_INCLUDE__
-#include "minocore/bicriteria.h"
-#include "minocore/blaze_adaptor.h"
-#include "minocore/Inf2Zero.h"
-#include "minocore/caratheodory.h"
-#include "minocore/coreset.h"
-#include "minocore/csc.h"
-#include "minocore/diskmat.h"
-#include "minocore/distance.h"
-#include "minocore/div.h"
-#include "minocore/minocore.h"
-#include "minocore/geo.h"
-#include "minocore/gmm.h"
-#include "minocore/graph.h"
-#include "minocore/hash.h"
-#include "minocore/applicator.h"
-#include "minocore/jv.h"
-#include "minocore/jv_solver.h"
-#include "minocore/kcenter.h"
-#include "minocore/kmeans.h"
-#include "minocore/kmedian.h"
-#include "minocore/lsearch.h"
-#include "minocore/macros.h"
-#include "minocore/matrix_coreset.h"
-#include "minocore/parse.h"
-#include "minocore/poisson_diss.h"
-#include "minocore/shared.h"
-#include "minocore/timer.h"
+#include <minocore/utility.h>
+
+#include <minocore/dist.h>
+#include <minocore/coreset.h>
+
+#include <minocore/optim.h>
+
+#include <minocore/wip.h>
+
 #endif
