+++ include/minocore/optim.h
@@ -1,0 +1,15 @@
+#ifndef FGC_OPTIM_H__
+#define FGC_OPTIM_H__
+
+// Continuous solvers
+#include <optim/kcenter.h>
+#include <optim/kmeans.h>
+#include <optim/kmedian.h>
+
+// Discrete solvers
+#include <optim/bicriteria.h>
+#include <optim/jv_solver.h>
+#include <optim/jv.h>
+#include <optim/lsearch.h>
+
+#endif
