+++ include/minocore/optim/kmeans.h
@@ -4,9 +4,9 @@
 #include <cassert>
 #include <mutex>
 #include <numeric>
-#include "matrix_coreset.h"
-#include "timer.h"
-#include "div.h"
+#include "minocore/coreset/matrix_coreset.h"
+#include "minocore/util/timer.h"
+#include "minocore/util/div.h"
 
 namespace minocore {
 using blz::rowiterator;
@@ -236,11 +236,11 @@ kmc2(const Oracle &oracle, RNG &rng, size_t np, size_t k, size_t m = 2000)
 {
     if(m == 0) throw std::invalid_argument("m must be nonzero");
     schism::Schismatic<IT> div(np);
-    flat_hash_set<IT> centers{div.mod(IT(rng()))};
+    shared::flat_hash_set<IT> centers{div.mod(IT(rng()))};
     if(*centers.begin() > np) throw std::runtime_error("ZOMGSDFD");
     // Helper function for minimum distance
     auto mindist = [&centers,&oracle](auto newind) {
-        typename flat_hash_set<IT>::const_iterator it = centers.begin(), end = centers.end();
+        typename shared::flat_hash_set<IT>::const_iterator it = centers.begin(), end = centers.end();
         assert(centers.size());
         auto dist = oracle(*it, newind);
         while(++it != end) {
@@ -325,13 +325,12 @@ kmc2(const blaze::Matrix<MT, SO> &mat, RNG &rng, size_t k,
     return ret;
 }
 
-#define DUMBAVE 1
-
 template<typename IT, typename MatrixType, typename CMatrixType=MatrixType, typename WFT=double, typename Functor=blz::sqrL2Norm>
 double lloyd_iteration(std::vector<IT> &assignments, std::vector<WFT> &counts,
                        CMatrixType &centers, MatrixType &data,
                        const Functor &func=Functor(),
-                       const WFT *weights=nullptr)
+                       const WFT *weights=nullptr,
+                       bool use_moving_average=false)
 {
     static_assert(std::is_floating_point_v<WFT>, "WTF must be floating point for weighted kmeans");
     // make sure this is only rowwise/rowMajor
@@ -344,38 +343,63 @@ double lloyd_iteration(std::vector<IT> &assignments, std::vector<WFT> &counts,
     };
     OMP_ONLY(std::unique_ptr<std::mutex[]> mutexes = std::make_unique<std::mutex[]>(centers.rows());)
     centers = static_cast<typename CMatrixType::ElementType>(0.);
-    std::memset(counts.data(), 0, counts.size() * sizeof(counts[0]));
+    std::fill(counts.data(), counts.data() + counts.size(), WFT(0.));
     assert(blz::sum(centers) == 0.);
     bool centers_reassigned;
     std::unique_ptr<typename MatrixType::ElementType[]> costs;
     get_assignment_counts:
     centers_reassigned = false;
-
-    OMP_PRAGMA("omp parallel for schedule(dynamic)")
-    for(size_t i = 0; i < nr; ++i) {
-        assert(assignments[i] < centers.rows());
-        auto asn = assignments[i];
-        auto dr = row(data, i BLAZE_CHECK_DEBUG);
-        auto cr = row(centers, asn BLAZE_CHECK_DEBUG);
-        const auto w = getw(i);
-        {
-            OMP_ONLY(std::lock_guard<std::mutex> lg(mutexes[asn]);)
-            if(w == 1.) {
-                blz::serial(cr.operator+=(dr));
+    /*
+     *
+     * The moving average is supposed to be 
+     */
+    if(!use_moving_average) {
+        OMP_PRAGMA("omp parallel for schedule(dynamic)")
+        for(size_t i = 0; i < nr; ++i) {
+            assert(assignments[i] < centers.rows());
+            auto asn = assignments[i];
+            auto dr = row(data, i BLAZE_CHECK_DEBUG);
+            auto cr = row(centers, asn BLAZE_CHECK_DEBUG);
+            const auto w = getw(i);
+            {
+                OMP_ONLY(std::lock_guard<std::mutex> lg(mutexes[asn]);)
+                if(w == 1.) {
+                    blz::serial(cr.operator+=(dr));
+                }
+                else blz::serial(cr.operator+=(dr * w));
             }
-            else blz::serial(cr.operator+=(dr * w));
+            OMP_ATOMIC
+            counts[asn] += w;
         }
-        OMP_ATOMIC
-        counts[asn] += w;
+        OMP_PFOR
+        for(size_t i = 0; i < centers.rows(); ++i)
+            row(centers, i BLAZE_CHECK_DEBUG) *= (1. / counts[i]);
+    } else {
+        OMP_PRAGMA("omp parallel for schedule(dynamic)")
+        for(size_t i = 0; i < nr; ++i) {
+            assert(assignments[i] < centers.rows());
+            auto asn = assignments[i];
+            auto dr = row(data, i BLAZE_CHECK_DEBUG);
+            auto cr = row(centers, asn BLAZE_CHECK_DEBUG);
+            const auto w = getw(i);
+            {
+                OMP_ONLY(std::lock_guard<std::mutex> lg(mutexes[asn]);)
+                auto oldw = counts[asn];
+                if(!oldw) {
+                    cr = dr;
+                } else {
+                    cr += (dr - cr) * (w / (oldw + w));
+                }
+                counts[asn] = oldw + w;
+            }
+        }
     }
+#ifndef NDEBUG
     std::fprintf(stderr, "Assigned cluster centers\n");
+#endif
     for(size_t i = 0; i < centers.rows(); ++i) {
         VERBOSE_ONLY(std::fprintf(stderr, "center %zu has count %g\n", i, counts[i]);)
-        if(counts[i]) {
-#ifdef DUMBAVE
-            row(centers, i BLAZE_CHECK_DEBUG) *= (1. / counts[i]);
-#endif
-        } else {
+        if(!counts[i]) {
             if(!costs) {
                 std::srand(std::time(nullptr));
                 costs.reset(new typename MatrixType::ElementType[nr]);
@@ -400,7 +424,6 @@ double lloyd_iteration(std::vector<IT> &assignments, std::vector<WFT> &counts,
     }
     if(centers_reassigned)
         goto get_assignment_counts;
-    std::fprintf(stderr, "Assigning to centers\n");
     // 2. Assign centers
     double total_loss = 0.;
     OMP_PRAGMA("omp parallel for reduction(+:total_loss)")
@@ -433,14 +456,15 @@ double lloyd_loop(std::vector<IT> &assignments, std::vector<WFT> &counts,
                 CMatrixType &centers, MatrixType &data,
                 double tolerance=0., size_t maxiter=-1,
                 const Functor &func=Functor(),
-                const WFT *weights=nullptr)
+                const WFT *weights=nullptr,
+                bool use_moving_average=false)
 {
     if(tolerance < 0.) throw 1;
     size_t iternum = 0;
     double oldloss = std::numeric_limits<double>::max(), newloss;
     for(;;) {
         std::fprintf(stderr, "Starting iter %zu\n", iternum);
-        newloss = lloyd_iteration(assignments, counts, centers, data, func, weights);
+        newloss = lloyd_iteration(assignments, counts, centers, data, func, weights, use_moving_average);
         double change_in_cost = std::abs(oldloss - newloss) / std::min(oldloss, newloss);
         if(iternum++ == maxiter || change_in_cost <= tolerance) {
             std::fprintf(stderr, "Change in cost from %g to %g is %g\n", oldloss, newloss, change_in_cost);
