+++ include/minocore/util/blaze_adaptor.h
