+++ include/minocore/util/diskmat.h
