+++ include/minocore/util/oracle.h
@@ -2,7 +2,9 @@
 #define FGC_ORACLE_H__
 #include <vector>
 #include <mutex>
+#include <shared_mutex>
 #include <unordered_map>
+#include "./macros.h"
 
 namespace minocore {
 
@@ -46,12 +48,27 @@ make_oracle_wrapper(const Oracle &o, It start, It end) {
 
 template<typename IT>
 struct PairKeyType {
-    using Type = std::conditional_t<sizeof(IT) == 4, uint64_t, std::pair<IT, IT>>;
+    using Type = std::conditional_t<sizeof(IT) <= 4, uint64_t, std::pair<IT, IT>>;
     static constexpr Type make_key(IT lh, IT rh) {
         if constexpr(sizeof(IT) == 4) {
             return (uint64_t(lh) << 32) | rh;
         } else {
             return Type{lh, rh};
+        }
+    }
+    static auto lh(Type v) {
+        if constexpr(sizeof(IT) == 4) {
+            return uint32_t(v >> 32);
+        } else {
+            return v.first;
+        }
+    }
+    static auto rh(Type v) {
+        if constexpr(sizeof(IT) == 4) {
+            static constexpr Type bitmask = static_cast<IT>((uint64_t(1) << 32) - 1);
+            return v & bitmask;
+        } else {
+            return v.second;
         }
     }
 };
@@ -62,23 +79,35 @@ struct CachingOracleWrapper {
     using KeyType = PairKeyType<IT>;
     const Oracle &oracle_;
     Map<typename KeyType::Type, output_type> map_;
-    std::mutex mut_;
+private:
+    std::shared_mutex mut_;
+    using iterator_type = decltype(map_.find(std::declval<typename KeyType::Type>()));
+    // TODO: use two kinds of locks
+public:
     CachingOracleWrapper(const Oracle &oracle): oracle_(oracle) {}
     output_type operator()(IT lh, IT rh) {
         if constexpr(symmetric) {
             if(lh > rh) std::swap(lh, rh);
         }
-        auto key = make_key(lh, rh);
+        auto key = KeyType::make_key(lh, rh);
         output_type ret;
-        auto it = map_.find(key);
-        if(it == map_.end()) {
-            ret = oracle_(lh, rh);
-            if constexpr(threadsafe)
-            {
-                std::lock_guard<std::mutex> mut_;
-                map_.emplace(key, ret);
-            } else map_.emplace(key, ret);
-        } else ret = it->second;
+        {
+            std::shared_lock<decltype(mut_)> shared(mut_);
+            if(iterator_type it; (it = map_.find(key)) == map_.end()) {
+                ret = oracle_(lh, rh);
+                if constexpr(threadsafe)
+                {
+                    shared.unlock();
+                    std::unique_lock<decltype(mut_)> lock(mut_);
+                    map_.emplace(key, ret);
+                } else {
+                    map_.emplace(key, ret);
+                }
+            } else {
+                ret = it->second;
+                shared.unlock();
+            }
+        }
         return ret;
     }
 };
