+++ include/minocore/wip.h
@@ -1,0 +1,9 @@
+#ifndef FGC_WIP_H__
+#define FGC_WIP_H__
+
+#include "./wip/caratheodory.h"
+#include "./wip/streaming.h"
+#include "./wip/clustering.h"
+#include "./wip/gen_kmedian.h"
+
+#endif
